Fixed final bugs - 2 ( Added proper pdf file handling for api 29+ and students' Numbers data type changed to String)

+++ app/src/main/java/com/example/feesapplication/fragments/AddStudentFragment.kt
@@ -487,7 +487,7 @@ class AddStudentFragment : Fragment() {
             // Green signal to save data into database
             val newStudentData = Student(
                 studentName = studentNameField,
-                studentNumber = contactNumber.toLong(),
+                studentNumber = contactNumber,
                 feesAmount = feesField.toLong(),
                 feesStatus = sharedViewModel.parseFeesStatus(feeStatusSaved),
                 studentEmail = emailField,
